Update packstack-answers.success.txt
A few changes out of my head. After succesfully creating the bridged interfaces by hand and restarting/rebooting the server, this answer file should give you a working installation. Additionally the network set-up has to be done through the gui to be able to give floating ip addresses to the VMs.